Feat/loading (#8)
* done loading

* add loading
