+++ src/components/loading/index.js
@@ -1,0 +1,3 @@
+import Loading from './loading';
+
+export default Loading;
