+++ src/locales/en/translation.json
@@ -33,5 +33,6 @@
     "github": "",
     "instagram": "",
     "whatsapp": "",
-    "descritpion": "description"
+    "descritpion": "description",
+    "loading": "Loading"
 }
