+++ src/views/home/home.tsx
@@ -1,36 +1,44 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import Welcome from './components/welcome';
 import Button from '../../components/button';
+import Loading from '../../components/loading';
 import { useTranslation } from 'react-i18next';
 
 import logo from '../../images/logo.png';
 
 const Home: React.FC = () => {
     const { t } = useTranslation();
+    const [isLoading, setIsLoading] = useState(true);
 
     useEffect(() => {
-        const homeComponent = document.getElementById('home');
-        if (homeComponent) {
-            homeComponent.style.opacity = '1';
-        }
-    });
+        setTimeout(() => {
+            setIsLoading(false);
+            const homeComponent = document.getElementById('home');
+            if (homeComponent) {
+                homeComponent.style.opacity = '1';
+            }
+        }, 3500);
+    }, [isLoading]);
 
     return(
-        <div id="home" className="container">
-            <div className="row">
-                <div className="content-left">
-                    <Welcome />
-                    <div className="btn-wrapper">
-                        <Button className="contact-btn" label={t('contact_me')} />
-                    </div>    
-                </div>
-                <div className="content-right">
-                    <div className="logo-wrapper">
-                        <img src={logo} alt="logo" />
+        isLoading ?
+                    <Loading />
+                  :
+                    <div id="home" className="container">
+                        <div className="row">
+                            <div className="content-left">
+                                <Welcome />
+                                <div className="btn-wrapper">
+                                    <Button className="contact-btn" label={t('contact_me')} />
+                                </div>    
+                            </div>
+                            <div className="content-right">
+                                <div className="logo-wrapper">
+                                    <img src={logo} alt="logo" />
+                                </div>
+                            </div>
+                        </div>  
                     </div>
-                </div>
-            </div>  
-        </div>
     );
 };
