+++ src/views/skills/index.js
@@ -1,0 +1,3 @@
+import Skills from './skills';
+
+export default Skills;
