--- src/views/skills/skills.js
@@ -1,3 +1,0 @@
-import Skills from './index';
-
-export default Skills;
